requirements.txt
eth-brownie>=1.12.2,<2.0.0